breeze final

+++ app/Http/Controllers/ReportController.php
@@ -12,6 +12,7 @@ class ReportController extends Controller
     //
     public function index(Request $request)
     {
+
         //$reports= Report::all(); //выборка всех данных из таблицы reports
         /*$sort= $request->input('sort');
         if ($sort == 'asc' || $sort == 'desc')
@@ -46,6 +47,7 @@ class ReportController extends Controller
         }
 
         $status = $request->input('status');
+       
         $validate = $request->validate([
             'status' => "exists:statuses,id"
         ]);
@@ -53,12 +55,14 @@ class ReportController extends Controller
         if($validate)
         {
             $reports = Report::where('status_id', $status)
+                ->where('user_id', Auth::user()->id)
                 ->orderBy('created_at', $sort)
                 ->paginate(8);
         }
         else
         {
-            $reports = Report::orderBy('created_at', $sort)
+            $reports = Report::where('user_id', Auth::user()->id)
+                ->orderBy('created_at', $sort)
                 ->paginate(8);
         }
 
@@ -83,23 +87,56 @@ class ReportController extends Controller
 
     public function edit(Request $request, Report $report)
     {
-        return view('reports.show', compact('report'));
+        if (Auth::user()->id === $report->user_id)
+        {
+            return view('reports.show', compact('report'));
+        }
+        else
+        {
+            abort(403, 'У вас нет прав на редактирование этой записи.');
+        }
     }
 
     public function update (Request $request, Report $report)
     {
-        $data = $request-> validate([
-            'number' => 'string',
-            'description' => 'string',
-        ]);
+        if (Auth::user()->id === $report->user_id)
+        {
+            $data = $request->validate([
+                'number' => 'string',
+                'description' => 'string',
+            ]);
 
-        $report->update($data);
-        return redirect()->back();
+            $report->update($data);
+            return redirect()->back();
+        }
+        else
+        {
+            abort(403, 'У вас нет прав на обновление этой записи.');
+        }
     }
 
     public function destroy(Report $report)
     {
-        $report->delete();
-        return redirect()->back();
+        if (Auth::user()->id === $report->user_id)
+        {
+            $report->delete();
+            return redirect()->back();
+        }
+        else
+        {
+            abort(403, 'У вас нет прав на удаление этой записи.');
+        }
+    }
+
+    public function show(Request $request, Report $report)
+    {
+        if (Auth::user()->id === $report->user_id)
+        {
+            return view('reports.show', compact('report'));
+        }
+        else
+        {
+            abort(403, 'У вас нет прав на просмотр этой записи.');
+        }
     }
 }
